feat: Add exercise solutions for Chapter 6
This commit adds a new notebook, exercise_solutionts.ipynb, which contains solutions to exercises in Chapter 6 of the book. The notebook covers topics such as the use of CNN models for X-ray images, different CNN architectures for different dataset sizes, properties of convolutional neural networks, and the application of CNNs in various domains. The solutions provided in this notebook can be used as a reference for understanding and implementing CNN models in practice.